Initial poc, passes test suite
A first go at bidi algorithm, untested and most likely broken

Removed bidi conversion from utf8ToW and made a separate utf8ToWLogicalToVisual function. Bidi fails testing

Testcase from icu biditest passes

Changed bidi options and defaults, added string comparison helper to get better failure messages on u16 and u32 strings

Possibly working mac normalization, needs more testing

Move settings filler to langinfo

Added another test for bidi logic, fixed bug when removing control characters, cleanup tests a bit

Some cleanup and linux build fixes

Removed failOnBadChar flag, added a few Try... functions in it's place where needed. Will likely not build on anything but win32 currently

Added doxygen for all functions

Changed parameters for logicalToVisual functions
Fixed win32 build, untested on other platforms

Added icu to unified depends build system

Always treat icu as cross compile for tool building and data-preparation

Fixed broken build and broken tests

Fixed failing tests
added ifdef for normalizeSystemSafe to have it compile on icu older than v49 except for darwin

Reworked conversion and merged it all into a single Convert function. Added support for NFC, NFD and NFD-MAC normalization

Renamed methods to follow naming standard and be consistent

Added Normalization functions

Added testing for normalization and fixed NFD-Mac

Added more tests from Karlson2k.
Added new Try methods

+++ xbmc/win32/stdio_utf8.cpp
@@ -24,23 +24,31 @@
 int remove_utf8(const char* __filename)
 {
   std::wstring filenameW;
-  g_charsetConverter.utf8ToW(__filename, filenameW, false);
+  if (!g_charsetConverter.Utf8ToWSystemSafe(__filename, filenameW))
+    return -1;
   return ::DeleteFileW(filenameW.c_str()) ? 0 : -1;
 }
 
 int rename_utf8(const char* __old, const char* __new)
 {
   std::wstring oldW, newW;
-  g_charsetConverter.utf8ToW(__old, oldW, false);
-  g_charsetConverter.utf8ToW(__new, newW, false);
-  return ::MoveFileW(oldW.c_str(), newW.c_str()) ? 0 : -1;
+  if (g_charsetConverter.Utf8ToWSystemSafe(__old, oldW) &&
+      g_charsetConverter.Utf8ToWSystemSafe(__new, newW))
+  {
+    return ::MoveFileW(oldW.c_str(), newW.c_str()) ? 0 : -1;
+  }
+  return -1;
 }
 
 FILE* fopen64_utf8(const char* __filename, const char* __modes)
 {
   std::wstring filenameW, modesW;
-  g_charsetConverter.utf8ToW(__filename, filenameW, false);
-  g_charsetConverter.utf8ToW(__modes, modesW, false);
-  return _wfopen(filenameW.c_str(), modesW.c_str());
+  if (g_charsetConverter.Utf8ToWSystemSafe(__filename, filenameW) &&
+      g_charsetConverter.Utf8ToWSystemSafe(__modes, modesW))
+  {
+    return _wfopen(filenameW.c_str(), modesW.c_str());
+  }
+  errno = EINVAL;
+  return NULL;
 }
